move 4337-passkeys-singleton-signer to 4337-passkeys
