+++ examples/4337-passkeys/src/components/SwitchNetwork.tsx
