+++ examples/4337-passkeys/src/hooks/useEntryPointAccountBalance.ts
