+++ examples/4337-passkeys/src/hooks/useNativeTokenBalance.ts
