+++ examples/4337-passkeys/src/logic/safeWalletApp.ts
