+++ examples/4337-passkeys/src/vite-env.d.ts
