+++ examples/4337-passkeys/vite.config.ts
