+++ modules/passkey/test/4337/SafeWebAuthnSharedSigner.spec.ts
@@ -1,22 +1,9 @@
-import { buildSignatureBytes } from '@safe-global/safe-4337/src/utils/execution'
-import {
-  buildPackedUserOperationFromSafeUserOperation,
-  buildSafeUserOpTransaction,
-  calculateSafeOperationHash,
-} from '@safe-global/safe-4337/src/utils/userOp'
-import { encodeMultiSendTransactions } from '@safe-global/safe-4337/test/utils/encoding'
 import { expect } from 'chai'
 import { deployments, ethers } from 'hardhat'
 
-import * as ERC1271 from '../../utils/erc1271'
-import {
-  DUMMY_AUTHENTICATOR_DATA,
-  base64UrlEncode,
-  decodePublicKey,
-  encodeWebAuthnSignature,
-  getSignatureBytes,
-} from '../../../src/utils/webauthn'
-import { WebAuthnCredentials, encodeWebAuthnSigningMessage } from '../../utils/webauthnShim'
+import * as ERC1271 from '../utils/erc1271'
+import { DUMMY_AUTHENTICATOR_DATA, base64UrlEncode, getSignatureBytes } from '../../src/utils/webauthn'
+import { encodeWebAuthnSigningMessage } from '../utils/webauthnShim'
 
 const SIGNER_MAPPING_SLOT = BigInt(ethers.id('SafeWebAuthnSharedSigner.signer')) - 1n
 
@@ -287,243 +274,6 @@ describe('SafeWebAuthnSharedSigner', () => {
 
       expect(await target['isValidSignature(bytes32,bytes)'](dataHash, signature)).to.equal('0x00000000')
       expect(await target['isValidSignature(bytes,bytes)'](data, signature)).to.equal('0x00000000')
-    })
-  })
-
-  describe('Safe4337Module', () => {
-    const setupTests = deployments.createFixture(async ({ deployments }) => {
-      const {
-        SafeModuleSetup,
-        SafeL2,
-        SafeProxyFactory,
-        MultiSend,
-        FCLP256Verifier,
-        Safe4337Module,
-        EntryPoint,
-        SafeWebAuthnSharedSigner,
-      } = await deployments.fixture()
-
-      const [user] = await ethers.getSigners()
-      const entryPoint = await ethers.getContractAt('IEntryPoint', EntryPoint.address)
-      const module = await ethers.getContractAt(Safe4337Module.abi, Safe4337Module.address)
-      const proxyFactory = await ethers.getContractAt(SafeProxyFactory.abi, SafeProxyFactory.address)
-      const multiSend = await ethers.getContractAt('MultiSend', MultiSend.address)
-      const safeModuleSetup = await ethers.getContractAt(SafeModuleSetup.abi, SafeModuleSetup.address)
-      const singleton = await ethers.getContractAt(SafeL2.abi, SafeL2.address)
-      const sharedSigner = await ethers.getContractAt('SafeWebAuthnSharedSigner', SafeWebAuthnSharedSigner.address)
-      const verifiers = BigInt(FCLP256Verifier.address)
-
-      const navigator = {
-        credentials: new WebAuthnCredentials(),
-      }
-
-      return {
-        user,
-        proxyFactory,
-        multiSend,
-        safeModuleSetup,
-        module,
-        entryPoint,
-        singleton,
-        sharedSigner,
-        verifiers,
-        navigator,
-      }
-    })
-
-    describe('executeUserOp - new account', () => {
-      it('should execute user operation', async () => {
-        const { user, proxyFactory, multiSend, safeModuleSetup, module, entryPoint, singleton, sharedSigner, navigator, verifiers } =
-          await setupTests()
-
-        const credential = navigator.credentials.create({
-          publicKey: {
-            rp: {
-              name: 'Safe',
-              id: 'safe.global',
-            },
-            user: {
-              id: ethers.getBytes(ethers.id('chucknorris')),
-              name: 'chucknorris',
-              displayName: 'Chuck Norris',
-            },
-            challenge: ethers.toBeArray(Date.now()),
-            pubKeyCredParams: [{ type: 'public-key', alg: -7 }],
-          },
-        })
-        const publicKey = decodePublicKey(credential.response)
-
-        const initializer = singleton.interface.encodeFunctionData('setup', [
-          [sharedSigner.target],
-          1,
-          multiSend.target,
-          multiSend.interface.encodeFunctionData('multiSend', [
-            encodeMultiSendTransactions([
-              {
-                op: 1 as const,
-                to: safeModuleSetup.target,
-                data: safeModuleSetup.interface.encodeFunctionData('enableModules', [[module.target]]),
-              },
-              {
-                op: 1 as const,
-                to: sharedSigner.target,
-                data: sharedSigner.interface.encodeFunctionData('configure', [{ ...publicKey, verifiers }]),
-              },
-            ]),
-          ]),
-          module.target,
-          ethers.ZeroAddress,
-          0,
-          ethers.ZeroAddress,
-        ])
-        const safeSalt = Date.now()
-        const safe = await proxyFactory.createProxyWithNonce.staticCall(singleton.target, initializer, safeSalt)
-
-        const safeOp = buildSafeUserOpTransaction(
-          safe,
-          user.address,
-          ethers.parseEther('0.5'),
-          '0x',
-          await entryPoint.getNonce(safe, 0),
-          await entryPoint.getAddress(),
-          false,
-          false,
-          {
-            initCode: ethers.solidityPacked(
-              ['address', 'bytes'],
-              [
-                proxyFactory.target,
-                proxyFactory.interface.encodeFunctionData('createProxyWithNonce', [singleton.target, initializer, safeSalt]),
-              ],
-            ),
-            verificationGasLimit: 700000,
-          },
-        )
-        const safeOpHash = await module.getOperationHash(
-          buildPackedUserOperationFromSafeUserOperation({
-            safeOp,
-            signature: '0x',
-          }),
-        )
-
-        const assertion = navigator.credentials.get({
-          publicKey: {
-            challenge: ethers.getBytes(safeOpHash),
-            rpId: 'safe.global',
-            allowCredentials: [{ type: 'public-key', id: new Uint8Array(credential.rawId) }],
-            userVerification: 'required',
-          },
-        })
-        const signature = buildSignatureBytes([
-          {
-            signer: sharedSigner.target as string,
-            data: encodeWebAuthnSignature(assertion.response),
-            dynamic: true,
-          },
-        ])
-
-        await user.sendTransaction({ to: safe, value: ethers.parseEther('1') }).then((tx) => tx.wait())
-        expect(await ethers.provider.getBalance(safe)).to.equal(ethers.parseEther('1'))
-        expect(await ethers.provider.getCode(safe)).to.equal('0x')
-        expect(await sharedSigner.getConfiguration(safe)).to.deep.equal([0n, 0n, 0n])
-
-        await entryPoint.handleOps([buildPackedUserOperationFromSafeUserOperation({ safeOp, signature })], user.address)
-
-        expect(await ethers.provider.getBalance(safe)).to.be.lessThanOrEqual(ethers.parseEther('0.5'))
-        expect(await ethers.provider.getCode(safe)).to.not.equal('0x')
-        expect(await sharedSigner.getConfiguration(safe)).to.deep.equal([publicKey.x, publicKey.y, verifiers])
-
-        const [implementation] = ethers.AbiCoder.defaultAbiCoder().decode(['address'], await ethers.provider.getStorage(safe, 0))
-        expect(implementation).to.equal(singleton.target)
-
-        const safeInstance = singleton.attach(safe) as typeof singleton
-        expect(await safeInstance.getOwners()).to.deep.equal([sharedSigner.target])
-      })
-    })
-
-    describe('executeUserOp - existing account', () => {
-      it('should execute user operation', async () => {
-        const { user, proxyFactory, multiSend, safeModuleSetup, module, entryPoint, singleton, sharedSigner, navigator, verifiers } =
-          await setupTests()
-
-        const credential = navigator.credentials.create({
-          publicKey: {
-            rp: {
-              name: 'Safe',
-              id: 'safe.global',
-            },
-            user: {
-              id: ethers.getBytes(ethers.id('chucknorris')),
-              name: 'chucknorris',
-              displayName: 'Chuck Norris',
-            },
-            challenge: ethers.toBeArray(Date.now()),
-            pubKeyCredParams: [{ type: 'public-key', alg: -7 }],
-          },
-        })
-        const publicKey = decodePublicKey(credential.response)
-
-        const initializer = singleton.interface.encodeFunctionData('setup', [
-          [sharedSigner.target],
-          1,
-          multiSend.target,
-          multiSend.interface.encodeFunctionData('multiSend', [
-            encodeMultiSendTransactions([
-              {
-                op: 1 as const,
-                to: safeModuleSetup.target,
-                data: safeModuleSetup.interface.encodeFunctionData('enableModules', [[module.target]]),
-              },
-              {
-                op: 1 as const,
-                to: sharedSigner.target,
-                data: sharedSigner.interface.encodeFunctionData('configure', [{ ...publicKey, verifiers }]),
-              },
-            ]),
-          ]),
-          module.target,
-          ethers.ZeroAddress,
-          0,
-          ethers.ZeroAddress,
-        ])
-        const safeSalt = Date.now()
-        const safe = await proxyFactory.createProxyWithNonce.staticCall(singleton, initializer, safeSalt)
-        await proxyFactory.createProxyWithNonce(singleton, initializer, safeSalt)
-
-        const safeOp = buildSafeUserOpTransaction(
-          safe,
-          user.address,
-          ethers.parseEther('0.5'),
-          '0x',
-          await entryPoint.getNonce(safe, 0),
-          await entryPoint.getAddress(),
-        )
-        const { chainId } = await ethers.provider.getNetwork()
-        const safeOpHash = calculateSafeOperationHash(await module.getAddress(), safeOp, chainId)
-        const assertion = navigator.credentials.get({
-          publicKey: {
-            challenge: ethers.getBytes(safeOpHash),
-            rpId: 'safe.global',
-            allowCredentials: [{ type: 'public-key', id: new Uint8Array(credential.rawId) }],
-            userVerification: 'required',
-          },
-        })
-        const signature = buildSignatureBytes([
-          {
-            signer: sharedSigner.target as string,
-            data: encodeWebAuthnSignature(assertion.response),
-            dynamic: true,
-          },
-        ])
-
-        await user.sendTransaction({ to: safe, value: ethers.parseEther('1') }).then((tx) => tx.wait())
-        expect(await ethers.provider.getBalance(safe)).to.equal(ethers.parseEther('1'))
-
-        const userOp = buildPackedUserOperationFromSafeUserOperation({ safeOp, signature })
-        await entryPoint.handleOps([userOp], user.address)
-
-        expect(await ethers.provider.getBalance(safe)).to.be.lessThanOrEqual(ethers.parseEther('0.5'))
-      })
     })
   })
 })
